Tidied up some warnings.

+++ share/src/bi/netcdf/InputNetCDFBuffer.cpp
@@ -113,14 +113,14 @@ void bi::InputNetCDFBuffer::map() {
   /* ns dimension */
   nsDim = nc_inq_dimid(ncid, "ns");
   if (nsDim >= 0) {
-    BI_ERROR_MSG(ns < nc_inq_dimlen(ncid, nsDim),
+    BI_ERROR_MSG(ns < (int)nc_inq_dimlen(ncid, nsDim),
         "Given index " << ns << " outside range of ns dimension");
   }
 
   /* np dimension */
   npDim = nc_inq_dimid(ncid, "np");
   if (npDim >= 0) {
-    BI_ERROR_MSG(np < 0 || np < nc_inq_dimlen(ncid, npDim),
+    BI_ERROR_MSG(np < 0 || np < (int)nc_inq_dimlen(ncid, npDim),
         "Given index " << np << " outside range of np dimension");
   }
